Move resource_name() to tools from utils
tools and utils are similar names, let's keep just one
