+++ cli/src/lib.rs
@@ -22,4 +22,3 @@ pub mod rollback;
 pub mod secret;
 pub mod stats;
 pub mod tools;
-pub mod utils;
